Completed android potoken minting
